feat: ads hydration (#3)
* feat: facade with hydration logic

* feat: utils for isomorphic base64

* feat: mappers for ad server

* feat: args converter and context discovery

* feat: facets adapter

* refactor: use new Buffer node API

* feat: add facade testing

* refactor: change getAds client signature

Remove publisherId from params object, now use it in a different
parameter. This is useful for removing the AdServer client get ads type,
since we can use directly the AdRequest type if we pass the publisher ID
in a different parameter.

* chore: code formatting

* feat: add tests for getSkuIds from mappers.ts

* fix: type on test

* fix: types and add brand_name to ad request param

* fix: optimize getAttributeFromFacets

* fix: include productClusterNames as facet tag

* fix: remove unnecessary filter

* fix: remove test for non applicable use case

We should not check if the ad contains product_sku, as this is
guaranteed by the type system.