refactor: change hydration signatures
Previously, ProductFetcher received ads as payload. Now it receives
Offer and we changed Offer to include product ID. This simplifies code
because product fetcher doesn't need full sponsored details, only ids.

Also, the function buildHydratedProduct was moved to a new mappers
inside hydration with the new buildOffers, that was created to simplify
mergeAdsWithProducts.