feat: allow custom fetcher (#5)
* feat: allow custom fetcher

Instead of requiring the use of the public IS endpoint uncached, now
it's possible to allow a custom fetcher while still using the code that
merges ads and products.

* feat: add hydration module

Adds the hydration module composed of the product fetcher and ads and
products merger.

* fix: pass identity to fetcher instead publisherId

* chore: lint

* fix: tests

* chore: lint

* refactor: extract buildHydratedProduct

* refactor: change hydration signatures

Previously, ProductFetcher received ads as payload. Now it receives
Offer and we changed Offer to include product ID. This simplifies code
because product fetcher doesn't need full sponsored details, only ids.

Also, the function buildHydratedProduct was moved to a new mappers
inside hydration with the new buildOffers, that was created to simplify
mergeAdsWithProducts.

* feat: export hydration strategy components

* feat: export some types

* fix: allow btoa/atob work in node

* fix: update Ads naming

* fix: reuse getOffer functionpick

* feat: include new ad types